[DEV-69] refactor: 파일명 수정
